requirements.txt
google-adk
pydantic
asyncio
google-genai
google-adk[eval]
